requirements.txt
# Cryptofeed API Service Dependencies

# Core cryptofeed library
cryptofeed>=2.4.0

# FastAPI Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary>=2.9.0
asyncpg>=0.28.0
alembic==1.13.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client for REST API
httpx==0.25.2
aiohttp>=3.8.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# WebSocket dependencies
python-socks>=2.7.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async support
aiofiles>=23.0.0

# JSON processing
ujson>=5.7.0

# Utilities
python-multipart==0.0.6
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0

# Logging
structlog>=23.0.0

# System monitoring
psutil>=5.9.0

# Scheduling
schedule>=1.2.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0